Improve compilation process
The current architecture offers the possibility to add more plugins
and extend the its functionality. This flexibility has to be supported
during the compilation process. This change modify the Makefile to find
plugins and compile them.

It also adds a start.sh script which allows to run MultiCloud services
without the need of packaging their binaries in a container.

Change-Id: I9271c51daa4fd9156717e37f1818886c72e36c27
Signed-off-by: Victor Morales <victor.morales@intel.com>
Issue-ID: MULTICLOUD-301