Remove NFD from list of AddOns
The Node Feature Discovery AddOn has been included as part of the
capabilities supported by the KRD but it has been detected some
issues when its services are trying to start. Given that this feature
is not necessary for this phase it can be excluded from the default
list.

The _install_crictl function is not consumed by any other method,
this also includes its removal.

Change-Id: I771a8865894038b91324a9cb3797c2197666f480
Signed-off-by: Victor Morales <victor.morales@intel.com>
Issue-ID: MULTICLOUD-301