Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

Almost there, I only have a question about the way that you implemented the unit tests for vnfd module. The idea to use that approach is to make the methods of the unit test self-contained, defining the mock data in the same method where it's expected.

I forgot to mention before but there is an script[1] that is used for functional tests for the API, so maybe you should consider it for the next commit.

[1] https://github.com/onap/multicloud-k8s/blob/master/vagrant/tests/plugin.sh

Patch-set: 6
Label: Code-Review=-1