Document All-in-One Bare-Metal provisioning
This document describe the proces to do an All-in-One deployment in
a Bare-Metal deployment, listing the Hardware and Software minimal
requirements, it also uses the aio.sh bash script for reducing the
number of manual steps.

Vagrantfile and installer.sh has been modified to disable functional
tests by default and includes the passwordless sudo verification.

Change-Id: Iec7b0bb37d2a46342e6b7e60bed37dbdf2019a10
Signed-off-by: Victor Morales <victor.morales@intel.com>
Issue-ID: MULTICLOUD-408