Update patch set 3
Patch Set 3:

bare_metal_provisioning.rst doesn't seem to contain info on how to be 'consumed by Virtual Machines'. Are we planning to add that later?

Can you describe the change to the Vagrantfile in the commit message? It seems to be related to the installer.sh but I don't see this change reflected in the commit message.

Thanks

Patch-set: 3