Update patch set 1
Patch Set 1:

> @Kiran,
 > 
 > I think that you forgot to fix this https://gerrit.onap.org/r/#/c/73205/1/src/k8splugin/api/defhandler_test.go@344

The invalid bundle definition check is being done in the definition code and not in the handler. The handler only checks the body and readability of body. Since we are mocking the definition code, calling the handler with an invalid bundle definition code would test the same code as calling with a valid bundle definition ID.
That is why I added the invalid bundle definition check in the definition_test.go

Patch-set: 1