Add support for helm templates
Add support for evaluating helm templates
The interface provides a way to get a map which
contains a mapping from kubernetes kind to the
corresponding yaml file that defines a resource
of that kind. This map is then provided to the
instantiation code at instantiation time to create
in the kubernetes cluster.
P5: Use filepath.join instead of strings.join
P9: rebase with new folder structure

Issue-ID: MULTICLOUD-291
Change-Id: Ie75f5c616cc0cdc3e0ace49ff2c2f6c356a4c0d1
Signed-off-by: Kiran Kamineni <kiran.k.kamineni@intel.com>