Add support for downloading content
Add support for downloading content for creating
the merged helm charts. This api will be used mainly
by the profile api to create a converged chart which
will then be created by the instantiation code
P2: Add unit tests for archive.go
    Add download method for profile.go
    Update comments for download method
P3: Add unit test for empty files
P4: Add unit tests for Download functions

Note: Unit tests for the newly added download methods
will be added in a seperate patch as there are some
changes needed in the mockdb which are extensive and will
pollute this patch.

Issue-ID: MULTICLOUD-291
Change-Id: I9779eaf95366f527f0360eaddea663722c13b196
Signed-off-by: Kiran Kamineni <kiran.k.kamineni@intel.com>