Add test helm chart for testing
Add a test helm chart for testing the end to end
flow. This chart brings up a service and a deployment.
The deployment brings up two containers, vault and consul.
Vault and Consul are brought up in dev mode.
The helm chart also has a subchart, common, which contains
templates that are used within the main chart.
Added an override deployment.yaml which creates only one
container instead of two

Issue-ID: MULTICLOUD-291
Change-Id: I7eb10c9f43f8a4e09938139b5e2edee5d59feb3b
Signed-off-by: Kiran Kamineni <kiran.k.kamineni@intel.com>