Add more modes to  KRD Vagrant network config
KRD assumes all networking interfaces of type nat.
This patch removes that assumption and gives user
more choices like routed network for vagrant
network. By default KRD assumes 10.0.0.0/24 network
per VM and is provided in the config file. With
this patch the name of network as krd is required.
Examples in cofig/sample in this patch.

Change-Id: Ib518277b0862af02b0476aa377222ecfdb4f3fc0
Signed-off-by: Ritu Sood <ritu.sood@intel.com>
Issue-ID: MULTICLOUD-301