Add more modes to  KRD Vagrant network config
KRD assumes all networking interfaces of type nat.
This patch removes that assumption and gives user
more choices like routed network for vagrant
network. By default KRD assumes 10.10.10.10/24
network per VM and is provided in the config file.
With this patch the name of network as krd is
required.
Examples in cofig/samples in this patch. In file
config/samples/pdf.yml.mini shows example of
adding new networks in subset of VM's.

Change-Id: Ib518277b0862af02b0476aa377222ecfdb4f3fc0
Signed-off-by: Ritu Sood <ritu.sood@intel.com>
Issue-ID: MULTICLOUD-301