Update patch set 11
Patch Set 11:

> > > (1 comment)
 > >
 > > Not really. I don't want the patches to be on hold. I intend to
 > > close the others pretty soon after this. I need this to be merged
 > > because I need some files from this patch for adding unit tests
 > > (helm charts). Instead of duplicating those charts, I intend to
 > use
 > > them from this patch.
 > 
 > But if I change those lines the script doesn't work. I won't break
 > anything but it will be useless until the other patches are merged.

The problem is that this patch also adds some collateral like the test helm charts and so on which can be used in the unit tests for the other patches. I would have preferred the plugin.sh and other code changes to be made in a seperate patch rather than this one.

Patch-set: 11