Make aio.sh suitable for CI
aio.sh script currently clones multicloud/k8s repo itself.
That is unnessary and makes it impossible to hook k8s repository
to CI system where commit changeset is used to define repository
files to be tested.

This commit will remove cloning of multicloud/k8s repository
and also move aio.sh to baremetal folder which seems to be
the intended place for it.

Also documentation updated based on change.

Issue-ID: MULTICLOUD-629

Change-Id: Ibc975de7290cd29f02dce84c78388cbd3f797431
Signed-off-by: Samuli Silvius <s.silvius@partner.samsung.com>