Update patch set 1
Patch Set 1:

Is there a plan to put all the charts as subcharts so that there is only one Helm chart
A: Yes the plan is to create a single helm chart to deploy vFirewall. But, when we create network CRDs that will alter network configuration and cause problems while creating subsequent ones. It is my understanding that we need to introduce delays while doing so. So the current plan of action is to have them run individually and then align them to have a single chart bring up all the pods.

Patch-set: 1