Update patch set 3
Patch Set 3:

> aio.sh refers to scripts that are in vagrant folder.

From documents and from folder structure I assumed this has been the intended place for the aio.sh script. Is it not?

Referring to script in vagrant folder cannot be avoided. If that is the issue, then let's move aio.sh back to vagrant folder where it refers to those scripts. But then the story does not hold as the readme.md in vagrant folder only talks about vagrant deployment and readme.md in baremetal folder talks about aio. If we move it back can we assume docs can be also updated? And what's the point of whole hosting_providers folder if all entry scripts are in same folder (vagrant)?
Sorry, I don't know the history.

Patch-set: 3
Label: Code-Review=0