vFW usecase update
Update vFw Uses case to use 1.5.0
artifacts. 1.3.1 artifacts were
removed from nexus this week. Changed
the container images to use scripts.
In the current sink container IP
addresses are built in the image and
can't be changed or parametrized. Also
this approach makes it easy to change
images as and when there are changes in
demo/vFw repo.

Change-Id: Ib1aae1816de866889e93d61a77832680764acef6
Issue-ID: MULTICLOUD-623
Signed-off-by: Ritu Sood <ritu.sood@intel.com>