Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Which CI are we targeting here? ONAP CI is based on VMs, which means that nested VMs must already be supported on the host itself before we can bring up any VM in the CI job.
> For CI specifically, we need something that can be brought up in a VM via containers only. Eg: KIND or something like that.

We are targeting ONAP CI (https://gerrit.onap.org/r/c/ci-management/+/88103/).
If CI does not provide nested virtualization that would be a blocker. We did not even though about it wouldn't.

Minikube or KIND could be an alternative but then the KUD deployment itself would not get tested. But is it really Kubespray requiring nested virtualization? I though it's some of the demos (virtlet VMs)? As we would run this as aio not with vagrant.

Patch-set: 4
Reviewer: Gerrit User 3957 <3957@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0