Update patch set 1
Patch Set 1:

I am not sure I like the idea of running any kubectl commands on the master.
The idea of KUD is to install kubernetes on a bunch of nodes from an installer/handler node.
Kubectl should be available on the installer node if it is needed to execute any installations on kubernetes.
I am okay with removing the kubectl dependency, however, I would prefer if kubectl were installed in one place and be used during and after installation from the same location.
I am also not clear on why it should be run as root. Is it it because some other collateral requires it to be like that?

Patch-set: 1