Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > I am not sure I like the idea of running any kubectl commands on
>  > the master.
>  > The idea of KUD is to install kubernetes on a bunch of nodes from
>  > an installer/handler node.
>  > Kubectl should be available on the installer node if it is needed
>  > to execute any installations on kubernetes.
>  > I am okay with removing the kubectl dependency, however, I would
>  > prefer if kubectl were installed in one place and be used during
>  > and after installation from the same location.
>  > I am also not clear on why it should be run as root. Is it it
>  > because some other collateral requires it to be like that?
> 
> Spoke to Kiran. Running kubectl within the cluster should be fine as this is during installation. The why root question needs clarification

The patch to whihc this patch is based on, removed kubectl installation (and causes the root problem). This patch fixes it by using kubectl on master as it's there by default installed by kubespray. End result is that there is kubectl installation just ones (on master) not two times (also on installer) and it's always used from master.

Running as root is because installer.sh installs the whole kubespray as root and kubectl on master also configured for root. Root user should be used when using kubectl on master.

Alternatively we could install kubectl also on installer node (but that you already removed in previous patch). But as your configuration is copying kubectl configs usable on installer
https://gerrit.onap.org/r/gitweb?p=multicloud/k8s.git;a=blob;f=kud/hosting_providers/vagrant/inventory/group_vars/k8s-cluster.yml;h=8f719a430aa49a45604bf0fcb4009112c06ce48c;hb=HEAD#l42
it's possible to configure kubectl on installer.

Installer.sh runs the addons ansible play as non-root and it's possible to install something as non-root (vagrant) user so we could return back the kubectl installation to installer node  (as non-root) but that will break aio installation as installer node and master is same in that case and that was the problem where all started (i.e. conflicting kubectl versions coming from kubespray and the one we install). And if you try aio you notice you need to use kubectl as root user as that's the kubectl from kubespray. But as aio is used as root it's not noticed.

I can think of a solution where we install own kubectl on installer in multinode case but not install it on the case where installer node = master.

Patch-set: 1
Reviewer: Gerrit User 3957 <3957@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0