Update patch set 3
Patch Set 3:

Hi,
Yes, this change was tested both with and without ovn-kubernetes addon installed (and worked in both scenarios).
I was not aware of lack of support for ovn-kubernetes in multicloud-k8s' KUD, but, nevertheless, I think this fix could still stay, as it's flexible enough to work in both environments.
Do you agree?

Patch-set: 3