Use Multus Daemonset for installing Multus in KuD
Currently KuD uses Ansible scripts for installing Multus.
Multus has a daemonset that should be used for installing the multus
as part of an add-on.
This is also helpful for KuD offline deployment in the future.
Removed the comment. Updated the images path and removed
error supression addressed by comments.

Signed-off-by: Akhila Kishore <akhila.kishore@intel.com>
Issue-ID: MULTICLOUD-681
Change-Id: Id3702a2b5bd18804c2d7e4d063eba656202cb840