Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> how to test it?

Hi Haibin, since these are just interfaces without implementation code, testing would be moot. The purpose of this patch was to have a collective conversation and agree on how the hpa plugin and scheduler would talk. Implementation code and testing code is intended to be added in a later patch. However, as of now, Srini+Ritu changed the design, so that hpa placement would be done by a separate micro-service using gRPC. This patch might be abandoned as such.

Patch-set: 14