Update patch set 2
Patch Set 2:

All of the API patches are going to conflict at api.go, projectHandler.go, main.go if they are designed/patterned after the current project API.

I think it's bad design when code is adding arguments to functions to ease testing, this is one of the reasons this code conflicts. Design of testing should be separate and not materially affect the design of production code.

Patch-set: 2