Update patch set 5
Patch Set 5: Code-Review-2

Discussed this patch with Ritu. We agreed to change the name of the API to controllers.

Also storeName should be changed to collectionName and set to 'controllers'. This collectionName should be used to create new controllers and to read from the controllers. All of the controller API DB storage should be contained in controllers collection in mongoDB. We should not be creating a new collection per create as was done in project API.

Patch will be updated shortly.

Patch-set: 5
Label: Code-Review=-2