Update patch set 4
Patch Set 4: Code-Review-1

(3 comments)

Would it make sense for the dcm to follow ./pkg/infra/... and ./pkg/module/. structure of the orchestrator microservice?  (instead of the src/dcm/internal approach).

Just asking - in case the topic has not been discussed yet and the code was following the structure of hte k8splugin instead of the orchestrator code.

Patch-set: 4
Label: Code-Review=-1