Adding query APIs for AppIntents
In this patch the following tasks where performed:
1. Added APIs for query Intents for each App
in the compositeApp. For example, you can see all
intents of collectd or prometheus seperately if
both formed a compositeApp Collection.

2. Added 'provider-name' parameter to denote the cluster
objects used in all intents. Earlier we had only
clusterName and clusterlabel to denote a cluster.
Modified the tests also for this.

3. Fixed bugs in the plugin_collection_v2.sh.

4. Fixed some minor logging and formatting bugs.

Issue-ID: MULTICLOUD-1048
Signed-off-by: Rajamohan Raj <rajamohan.raj@intel.com>
Change-Id: Ic452d7ba5d98bb265eb301de84d679d1abe0e34e