Remove orchestrator dependency from ncm
NCM shall be like a library which orchestrator shall
call as and when required. NCM will assume that whenever
orchestrator calls it all the validation of orchestrator components
is done, so NCM shall not validate orchestrator components.
Otherwise, if NCM calls the orchestration, this will result
in cyclic import issue.

Issue-ID: MULTICLOUD-1041
Signed-off-by: Rajamohan Raj <rajamohan.raj@intel.com>
Change-Id: Ie352c730ffd39863d6de23d6a9a9a934f460392b