Update patch set 7
Patch Set 7: Code-Review-1

For the resource names in the app context, I have been working on the assumption that it will be of the form <name>+<kind> .
I've been following that approach in a couple patches so far.  If that is the general plan, then we should do it here.

Patch-set: 7
Reviewer: Gerrit User 1603 <1603@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1