Update patch set 2
Patch Set 2: Code-Review-1

Regarding adding of logging to context lib, it is not a good practice to log sensitive data. The handle is the key to a value pair in etcd. And also, this logs are in the normal execution path and they will fill up the trace/log buffers.

Patch-set: 2
Reviewer: Gerrit User 3087 <3087@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1