Temporarily comment-out monitor module build in src/Makefile
Also removes 'replace' directives in monitor/go.mod that cause
the build to fail with Go 1.14.4

The CI build currently fails with the change to monitor/go.mod
that is why the src/Makefile has the monitor build commented-out.

When the monitor build failure is resolved, the commented-out
monitor build can be restored.

Issue-ID: MULTICLOUD-1106
Signed-off-by: Larry Sachs <larry.j.sachs@intel.com>
Change-Id: Ic463129a0a608b4be74245ad4abe89c8c31bfd66