Create patch set 5
Uploaded patch set 5: Commit message was updated.

Patch-set: 5
Subject: This is a series of negative tests that validate the EMCO open api's. Description: In this series of tests, null is provided as an input to the POST, DELETE, and GET commands and the behavior is observed. In addition, the resultant error code is then captured and checked against current documentation to see whether or not the test failed as expected. Tests that failed to match the error code of the documentation are marked failed and those that match the error code are marked with success.
Commit: 28a69e06742e4c91a9df1134bbf074a35beea31c
Tag: autogenerated:gerrit:newPatchSet
Groups: 28a69e06742e4c91a9df1134bbf074a35beea31c