Update patch set 1
Patch Set 1:

Not saying the Postman request should necessarily be changed. But we should decide about whether it should or whether [1] should be adapted to not expect this top-level dict "metadata" (for cluster-provider it's obviously adequate).

Also, looking at other metadata uses maybe this is also a problem in other resources like composite app? Need to make it all consistent.

[1] https://github.com/onap/multicloud-k8s/blob/39b7d2f8a79f23141301e7b9abf7bf3a0f2e8f21/src/clm/api/clusterhandler.go#L140

Patch-set: 1