Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> Move the helm chart under kud directory.

The other Helm charts in this directory are related to installing onap4k8s/k8splugin. Some other helm charts here are supposed to work with k8splugin. Since the source code for the SDWAN application is not in this repo this Helm chart is related to use case for k8splugin. Only this chart needs to be moved.

Patch-set: 3
Reviewer: Gerrit User 285 <285@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>