Reimplement Terminate to be compatible with Status
This also includes modifying Create, Apply and Delete, since there are
strict conditions that need to be met in each, to prevent
Logical Clouds (LCs) from entering a bad state.

Summary of what's being done here:
- When creating:
  - set LC 'terminating' flag to false - because it's not terminating
- When applying:
  - set tag 'lccontext' in the LC to the context ID (was already done)
  - and let rsync know about the appcontext (grpc) (was already done)
  - if tag was already set, check current context /status
  - if context /status is actually terminated, both 'lccontext' and
'terminating' are reset (deleted and set to false respectively)
- When terminating:
  - simply flags the LC as 'terminating' (terminating=true)
  - and lets rsync know about the termination request (grpc)
- When deleting:
  - checks whether the current context /status is Terminated
  - if it is, then it will remove the LC context

The 'terminate' flag is simply used in multiple points to figure out
what state we're on without ever having to lose the 'lccontext' tag.
'lccontext' keeps track of the last context ID which is necessary
for DCM to know what is the current context /status.

This particular commit disables the TestDeleteLogicalCloud test
until a known issue behind the test is resolved.

This commit does not leverage the full capacity of the Status
framework, but is sufficient to support all operations. A future
patch will entirel migrate DCM to the Status framework.

Issue-ID: MULTICLOUD-1143
Change-Id: I7a6034eba543c2a27daa41b7fe6298cb2a85f9ce
Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>