Update multus-daemonset.yml to that of multus-cni v3.6 release
The only change to the upstream yml is the removal of the
kube-multus-ds-ppc64le DaemonSet and the replacement of
"default-cni-network" with "cni0".

Note also that the v3.6 yml actually uses the v3.4.1 image tag.  The
yml now points to a v3.4.1 image with the addition of code to merge
the results from all delegates to support Virtlet.

Issue-ID: MULTICLOUD-1230
Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>
Change-Id: I0e18644a567facfac1fd7dc1c053002b2d906288