Update patch set 1
Patch Set 1:

recheck

I've studied the failing log and tried to reproduce with no luck.  The log provides no information on what exactly failed.  Even more puzzling, the changes in this commit are not even executed in the verify builder (KUD_ADDONS is set to only virtlet and ovn4nfv: it would need to include optane to test this change).

Patch-set: 1