WIP Enable pod security policies
CURRENT STATE:
- the ovn4nfv and nfd addons pass
- virtlet was not tested, but it sits in the kube-system namespace so
  this should not be a problem
- sriov passed, but no SR-IOV hardware was present in the test
  environment
- cmk is failing due to errors in the nginx pod used in the test
  script
- emco deploys and emco test passes; plugin_fw_v2 test remains to be
  tested
- onap4k8s has not been tested

The intention with this change is to disable CAP_NET_RAW (which can be
a security vulnerability) for created Pods.

kubespray provides the podsecuritypolicy_enabled variable for enabling
privileged (for kube-system) and restricted (for everyone else)
policies.  Enabling this requires binding the KUD_ADDONs to the
privileged policy and specifying the security context correctly for
Pods running in the default namespace.

Issue-ID: MULTICLOUD-1256
Signed-off-by: Todd Malsbary <todd.malsbary@intel.com>
Change-Id: I7d6add122ad4046f9116ef03a249f5c9da1d7eec